Renamed files for clarity. From mainLayout to MainScreen
