+++ app/src/main/java/com/exercise/matipv2/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.exercise.matipv2.ui.MainLayout
+import com.exercise.matipv2.ui.MainScreen
 import com.exercise.matipv2.ui.theme.MaTIPv2Theme
 
 class MainActivity : ComponentActivity() {
@@ -22,7 +22,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    MainLayout(viewModel())
+                    MainScreen(viewModel())
                 }
             }
         }
@@ -33,6 +33,6 @@ class MainActivity : ComponentActivity() {
 @Composable
 fun MatipPreview() {
     MaTIPv2Theme {
-        MainLayout(viewModel())
+        MainScreen(viewModel())
     }
 }
