--- app/src/main/java/com/exercise/matipv2/model/MainLayoutState.kt
@@ -1,9 +1,0 @@
-package com.exercise.matipv2.model
-
-data class MainLayoutState(
-    var amountInput: String = "",
-    var tipPercentInput: String = "",
-    var roundUp: Boolean = false,
-    var splitShare: Int = 0,
-    var selectedTabIndex: Int = 0
-)
