+++ app/src/main/java/com/exercise/matipv2/model/MainScreenState.kt
@@ -1,0 +1,9 @@
+package com.exercise.matipv2.model
+
+data class MainScreenState(
+    var amountInput: String = "",
+    var tipPercentInput: String = "",
+    var roundUp: Boolean = false,
+    var splitShare: Int = 0,
+    var selectedTabIndex: Int = 0
+)
